+++ lib/jazz_fingers/prompt.rb
@@ -1,9 +1,14 @@
 # frozen_string_literal: true
 
 module JazzFingers
-  # For Pry >= 0.13.
   class Prompt
     OBJECT_INSTANCE = /#<(.+)>/
+
+    if Pry::VERSION >= "0.13.0"
+      include Pry013
+    else
+      include Pry012
+    end
 
     def initialize(options = {})
       @colored = options.fetch(:colored)
@@ -108,22 +113,6 @@ module JazzFingers
         line: line_number(pry),
         separator: separator
       )
-    end
-
-    # Add the JazzFingers prompt to the Pry::Prompt hash to enable changing it
-    # with `change-prompt`. Return the Pry::Prompt object.
-    def  pry_config
-      return Pry::Prompt[:jazz_fingers] if Pry::Prompt[:jazz_fingers]
-
-      Pry::Prompt.add(
-        :jazz_fingers,
-        "A spruced-up prompt provided by jazz_fingers.",
-        [main_separator, wait_separator]
-      ) do |context, _nesting, pry, separator|
-        template(Pry.view_clip(context), pry, separator)
-      end
-
-      Pry::Prompt[:jazz_fingers]
     end
   end
 end
